+++ packages/ui/src/molecules/InvoiceMetaDetails.tsx
@@ -1,3 +1,4 @@
+import { ExternalLinkIcon } from '@chakra-ui/icons';
 import {
   Button,
   Heading,
@@ -11,12 +12,11 @@ import {
   Wrap,
   WrapItem,
 } from '@chakra-ui/react';
+import { INVOICE_TYPES } from '@smartinvoicexyz/constants';
 import { InvoiceDetails } from '@smartinvoicexyz/types';
 import {
   chainByName,
   documentToHttp,
-  getAccountString,
-  getAddressLink,
   getDateString,
 } from '@smartinvoicexyz/utils';
 import _ from 'lodash';
@@ -28,6 +28,7 @@ import {
   AccountLink,
   CopyIcon,
   GenerateInvoicePDF,
+  InvoiceBadge,
   QuestionIcon,
   VerifyInvoice,
 } from '..';
@@ -48,6 +49,7 @@ export function InvoiceMetaDetails({
     provider,
     resolver,
     verified,
+    invoiceType,
     resolverInfo,
   } = _.pick(invoice, [
     'id',
@@ -58,6 +60,7 @@ export function InvoiceMetaDetails({
     'deadline',
     'metadata',
     'verified',
+    'invoiceType',
     'resolverInfo',
   ]);
 
@@ -108,7 +111,7 @@ export function InvoiceMetaDetails({
             Number(terminationTime) * 1000,
           ).toUTCString()}`,
         },
-      deadline && {
+      !!deadline && {
         label: 'Payment Deadline:',
         value: getDateString(_.toNumber(_.toString(deadline))),
         tip: `Late fees start accumulating after ${new Date(
@@ -165,34 +168,35 @@ export function InvoiceMetaDetails({
         )}
 
         <HStack align="center" color="black" spacing={4}>
-          <Link
-            href={getAddressLink(invoiceChainId, _.toLower(invoiceId))}
-            isExternal
-          >
-            {getAccountString(invoiceId as Address | undefined)}
-          </Link>
+          <InvoiceBadge invoiceType={invoiceType} />
+          <AccountLink
+            address={invoiceId as Address}
+            chainId={invoiceChainId}
+          />
           <Button
             onClick={onCopy}
             variant="ghost"
+            bg="none"
             colorScheme="blue"
             h="auto"
             w="auto"
             minW="2"
             p={2}
           >
-            <CopyIcon boxSize={4} />
+            <CopyIcon boxSize={3} />
           </Button>
         </HStack>
         {description && <Text color="black">{description}</Text>}
 
         {!!lastDocument && (
-          <Link
-            href={documentToHttp(lastDocument)}
-            isExternal
-            textDecor="underline"
-            color="black"
-          >
-            Details of Agreement
+          <Link href={documentToHttp(lastDocument)} isExternal _hover={{}}>
+            <Button
+              size="xs"
+              textTransform="uppercase"
+              rightIcon={<ExternalLinkIcon />}
+            >
+              View Details of Agreement
+            </Button>
           </Link>
         )}
       </Stack>
@@ -222,27 +226,29 @@ export function InvoiceMetaDetails({
           </Wrap>
         ))}
 
-        <Wrap>
-          <WrapItem>
-            <Text>{'Non-Client Deposits Enabled: '}</Text>
-          </WrapItem>
+        {invoiceType === INVOICE_TYPES.Escrow && (
+          <Wrap>
+            <WrapItem>
+              <Text>{'Non-Client Deposits Enabled: '}</Text>
+            </WrapItem>
 
-          <WrapItem fontWeight="bold">
-            {invoice && verifiedStatus ? (
-              <Text color="green.500">Enabled!</Text>
-            ) : (
-              <Text color="red.500">Not enabled</Text>
-            )}
-          </WrapItem>
+            <WrapItem fontWeight="bold">
+              {invoice && verifiedStatus ? (
+                <Text color="green.500">Enabled!</Text>
+              ) : (
+                <Text color="red.500">Not enabled</Text>
+              )}
+            </WrapItem>
 
-          <WrapItem fontWeight="bold">
-            <VerifyInvoice
-              invoice={invoice}
-              isClient={isClient}
-              verifiedStatus={verifiedStatus}
-            />
-          </WrapItem>
-        </Wrap>
+            <WrapItem fontWeight="bold">
+              <VerifyInvoice
+                invoice={invoice}
+                isClient={isClient}
+                verifiedStatus={verifiedStatus}
+              />
+            </WrapItem>
+          </Wrap>
+        )}
 
         <Wrap>
           <GenerateInvoicePDF