+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @smartinvoicexyz/hooks
 
+## 0.1.13
+
+### Patch Changes
+
+- refactored chains
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.8
+  - @smartinvoicexyz/graphql@0.1.10
+  - @smartinvoicexyz/types@0.1.10
+  - @smartinvoicexyz/utils@0.1.11
+
+## 0.1.12
+
+### Patch Changes
+
+- view invoices without wallet
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.7
+  - @smartinvoicexyz/graphql@0.1.9
+  - @smartinvoicexyz/types@0.1.9
+  - @smartinvoicexyz/utils@0.1.10
+
 ## 0.1.11
 
 ### Patch Changes