+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @smartinvoicexyz/shared
 
+## 0.1.8
+
+### Patch Changes
+
+- refactored chains
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.8
+
+## 0.1.7
+
+### Patch Changes
+
+- view invoices without wallet
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes