+++ packages/types/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @smartinvoicexyz/types
 
+## 0.1.10
+
+### Patch Changes
+
+- refactored chains
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.8
+  - @smartinvoicexyz/graphql@0.1.10
+  - @smartinvoicexyz/shared@0.1.8
+
+## 0.1.9
+
+### Patch Changes
+
+- view invoices without wallet
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.7
+  - @smartinvoicexyz/graphql@0.1.9
+  - @smartinvoicexyz/shared@0.1.7
+
 ## 0.1.8
 
 ### Patch Changes