+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @smartinvoicexyz/utils
 
+## 0.1.11
+
+### Patch Changes
+
+- refactored chains
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.8
+  - @smartinvoicexyz/graphql@0.1.10
+  - @smartinvoicexyz/shared@0.1.8
+  - @smartinvoicexyz/types@0.1.10
+
+## 0.1.10
+
+### Patch Changes
+
+- view invoices without wallet
+- Updated dependencies
+  - @smartinvoicexyz/constants@0.1.7
+  - @smartinvoicexyz/graphql@0.1.9
+  - @smartinvoicexyz/shared@0.1.7
+  - @smartinvoicexyz/types@0.1.9
+
 ## 0.1.9
 
 ### Patch Changes